Migrate cookbook to use jshell
