+++ java/source/demo/src/main/java/create/Util.java
@@ -1,4 +1,5 @@
-import org.apache.arrow.memory.RootAllocator;
+package create;
+
 import org.apache.arrow.vector.BitVectorHelper;
 import org.apache.arrow.vector.IntVector;
 import org.apache.arrow.vector.VarCharVector;
@@ -8,28 +9,8 @@ import org.apache.arrow.vector.types.Types;
 import org.apache.arrow.vector.types.pojo.FieldType;
 
 import java.util.List;
-import static java.util.Arrays.asList;
 
-public class Create {
-    public static void main(String[] args) {
-        RootAllocator rootAllocator = new RootAllocator(Long.MAX_VALUE);
-
-        // create int vector
-        IntVector intVector = new IntVector("intVector", rootAllocator);
-        setVector(intVector, 1,2,3);
-        System.out.println(intVector);
-
-        // create a varchar vector
-        VarCharVector varcharVector = new VarCharVector("varcharVector", rootAllocator);
-        setVector(varcharVector, "david".getBytes(), "gladis".getBytes(), "juan".getBytes());
-        System.out.println(varcharVector);
-
-        // create a list vector
-        ListVector listVector = ListVector.empty("listVector", rootAllocator);
-        setVector(listVector, asList(1,3,5,7,9), asList(2,4,6,8,10), asList(1,2,3,5,8));
-        System.out.println(listVector);
-    }
-
+public class Util {
     public static void setVector(IntVector vector, Integer... values) {
         final int length = values.length;
         vector.allocateNew(length);
